FTBbox now uses floats rather then ints

+++ include/FTFont.h
@@ -106,7 +106,7 @@ class FTGL_EXPORT FTFont
 		 * @param ury		upper right far y coord
 		 * @param urz		upper right far z coord
 		 */
-		void BBox( const char* text, int& llx, int& lly, int& llz, int& urx, int& ury, int& urz);
+		void BBox( const char* text, float& llx, float& lly, float& llz, float& urx, float& ury, float& urz);
 
 		/**
 		 * Get the bounding box for a string.
@@ -119,7 +119,7 @@ class FTGL_EXPORT FTFont
 		 * @param ury		upper right far y coord
 		 * @param urz		upper right far z coord
 		 */
-		void BBox( const wchar_t* string, int& llx, int& lly, int& llz, int& urx, int& ury, int& urz);
+		void BBox( const wchar_t* string, float& llx, float& lly, float& llz, float& urx, float& ury, float& urz);
 
 		/**
 		 * Get the advance width for a string.
